+++ include/arch/x86/arch.h
@@ -16,6 +16,7 @@
 #include <stdbool.h>
 #include <irq.h>
 #include <arch/x86/mmustructs.h>
+#include <arch/x86/thread_stack.h>
 
 static ALWAYS_INLINE void z_arch_irq_unlock(unsigned int key)
 {
