+++ include/arch/x86/intel64/arch.h
@@ -7,9 +7,7 @@
 #define ZEPHYR_INCLUDE_ARCH_X86_INTEL64_ARCH_H_
 
 #include <arch/x86/intel64/thread.h>
-
-#define STACK_ALIGN 16
-#define STACK_SIZE_ALIGN 16
+#include <arch/x86/thread_stack.h>
 
 #if CONFIG_ISR_STACK_SIZE != (CONFIG_ISR_SUBSTACK_SIZE * CONFIG_ISR_DEPTH)
 #error "Check ISR stack configuration (CONFIG_ISR_*)"
@@ -20,26 +18,6 @@
 #endif
 
 #ifndef _ASMLANGUAGE
-
-#define Z_ARCH_THREAD_STACK_LEN(size) (ROUND_UP((size), STACK_SIZE_ALIGN))
-
-#define Z_ARCH_THREAD_STACK_DEFINE(sym, size) \
-	struct _k_thread_stack_element __noinit \
-		__aligned(STACK_ALIGN) \
-		sym[Z_ARCH_THREAD_STACK_LEN(size)]
-
-#define Z_ARCH_THREAD_STACK_ARRAY_DEFINE(sym, nmemb, size) \
-	struct _k_thread_stack_element __noinit \
-		__aligned(STACK_ALIGN) \
-		sym[nmemb][Z_ARCH_THREAD_STACK_LEN(size)]
-
-#define Z_ARCH_THREAD_STACK_MEMBER(sym, size) \
-	struct _k_thread_stack_element __aligned(STACK_ALIGN) \
-		sym[Z_ARCH_THREAD_STACK_LEN(size)]
-
-#define Z_ARCH_THREAD_STACK_SIZEOF(sym)	sizeof(sym)
-#define Z_ARCH_THREAD_STACK_BUFFER(sym) ((char *) sym)
-
 static ALWAYS_INLINE unsigned int z_arch_irq_lock(void)
 {
 	unsigned long key;
